requirements.txt
flask==2.0.3
werkzeug==2.0.3
requests
python-dotenv
gunicorn
psycopg2-binary
